backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
httpx
beautifulsoup4
pydantic
python-multipart
openai
